initial commit
The initial commit consists entirely of the first README version.  Because what better
show of hubris than insisting on writing a README for all my future users and
contributors instead of actual code? The markdown honestly hasn't been tested, but the few uses
 of it I've made are fresh enough in my mind that I think I know what it'll look like.
Internet is down right now and I doubt this will gain any traction anyway.  If it does,
and you are reading this initial commit: HELLO FROM THE PAST (send pizza!)